Provide globalization member to ZoweZLUX global object
Provide an implementation  to the ZoweZLUX global object.

This is a fairly sensitive bootstrapping problem, and I think that
we will probably find a better way, perhaps involving moving the
underlying REST service from zlux/app-manager/virtual-desktop to
zlux-proxy-server as a "built-in" plugin.

Fortunately, the user of the global object should not be aware
of any future reorganization of the bootstrapping.

Signed-off-by: Robert Penny <rpenny@rocketsoftware.com>